backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
openai==1.30.1  # if you use Groq's OpenAI-compatible SDK
